Merges similar classes
