--- m3/Packages/com.oldbard.services.match3/Runtime/Grid/GridMatchesValidator.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: cec4cc72aa1147f4bfd1e066658be1f5
-timeCreated: 1709046726
